[ADD] 이거저거추가

+++ db/migrations/20220920025339_create_reviews_table.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS reviews (
   room_id INT NOT NULL,
   reservation_id INT NOT NULL,
   review varchar(1000) NOT NULL,
-  rating INT NOT NULL,
+  rating decimal(3,1) NOT NULL,
   created_at TIMESTAMP DEFAULT NOW(),
   updated_at TIMESTAMP DEFAULT NOW() ON UPDATE NOW(),
   PRIMARY KEY (id),
